requirements.txt
ffmpeg-python==0.2.0
pyyaml
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==2.3.2
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
requests==2.28.2
urllib3==1.26.20
Werkzeug==3.1.3
